+++ cmd/provisioner/storage/pv_utils.go
@@ -1,13 +1,14 @@
-package pvc
+package storage
 
 import (
+	"fmt"
 	"k8s-pv-provisioner/cmd/provisioner/config"
-	"k8s-pv-provisioner/cmd/provisioner/storage"
 	"path"
+	"strings"
 
 	core_v1 "k8s.io/api/core/v1"
 	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	
+	"k8s.io/klog"
 )
 
 type pvArguments struct {
@@ -18,11 +19,13 @@ type pvArguments struct {
 	assetPath     string
 }
 
-func preparePv(pvc *core_v1.PersistentVolumeClaim) (*core_v1.PersistentVolume, error) {
+/*PreparePV is function which creates storage asset(folder) and returns prepared PV structure to be created in cluster. Depending on presence colon sign in
+StorageAssetRoot field of currentStorageClass NFS or HostPath type of PV will be returned*/
+func PreparePV(pvc *core_v1.PersistentVolumeClaim) (*core_v1.PersistentVolume, error) {
 	currentStorageClass := appConfig.StorageClasses[*pvc.Spec.StorageClassName]
 
-	uid, gid := storage.ChooseAssetOwner(pvc)
-	storageAssetBaseName := storage.ChooseBaseNameOfAsset(pvc.Namespace, pvc.Name)
+	uid, gid := ChooseAssetOwner(pvc)
+	storageAssetBaseName := ChooseBaseNameOfAsset(pvc.Namespace, pvc.Name)
 
 	/*appStorageAssetPath is the full path to storage asset (folder) as it is seen or reachable from container of the provisioner*/
 	appStorageAssetPath := path.Join(appConfig.StorageAssetRoot, currentStorageClass.Name, storageAssetBaseName) // e.g. -> /pv-store/nfs-class1/sbx-namespace-some-app
@@ -33,7 +36,7 @@ func preparePv(pvc *core_v1.PersistentVolumeClaim) (*core_v1.PersistentVolume, e
 	if _, present := pvc.Annotations[config.AnnotationUseExistingAsset]; present {
 		reuseExistingAsset = true
 	}
-	err := storage.CreateStorageAsset(appStorageAssetPath, uid, gid, reuseExistingAsset)
+	err := CreateStorageAsset(appStorageAssetPath, uid, gid, reuseExistingAsset)
 	if err != nil {
 		return nil, err
 	}
@@ -56,13 +59,62 @@ func preparePv(pvc *core_v1.PersistentVolumeClaim) (*core_v1.PersistentVolume, e
 	pvArgs.reclaimPolicy = reclaimPolicy
 	pvArgs.pvc = pvc
 
-	pv := getHostPathPV(pvArgs)
+	pv := fillPV(pvArgs)
+
+	//This can happened after panic recovery
+	if pv == nil {
+		if !reuseExistingAsset {
+			/*If we are here and reuseExistingAsset == false than we can be sure that storage asset was created by us a few lines earlier.
+			On next iteration when the provioner will try to provision this PV it will face with issue that the storage asset already exists.
+			Therefore because of this issue we must delete created storage asset in current iteration when the panic occured*/
+			DeleteStorageAsset(appStorageAssetPath)
+		}
+
+		return pv, fmt.Errorf("Could not prepare new PV")
+	}
+
 	return pv, nil
 }
 
-func getHostPathPV(args *pvArguments) *core_v1.PersistentVolume {
+func getHostPathPersistentVolumeSource(assetPath string) core_v1.PersistentVolumeSource {
 	hostPathType := new(core_v1.HostPathType)
 	*hostPathType = core_v1.HostPathDirectory
+
+	return core_v1.PersistentVolumeSource{
+		HostPath: &core_v1.HostPathVolumeSource{
+			Type: hostPathType,
+			Path: assetPath,
+		},
+	}
+}
+func getNfsPersistentVolumeSource(assetPath string) core_v1.PersistentVolumeSource {
+	nfsAssetPath := strings.Split(assetPath, ":")
+	if len(nfsAssetPath) > 2 {
+		panic(fmt.Sprintf("A storage class assetRoot must contain only 1 colon sign if NFS-like path usage is assumed. Got value: %v", assetPath))
+	}
+	//TODO: Here should be some checks whether the nfsAssetPath[0] meet hostname reqirements. Implement them
+	return core_v1.PersistentVolumeSource{
+		NFS: &core_v1.NFSVolumeSource{
+			Server:   nfsAssetPath[0],
+			Path:     nfsAssetPath[1],
+			ReadOnly: false,
+		},
+	}
+}
+
+func fillPV(args *pvArguments) *core_v1.PersistentVolume {
+	defer func() {
+		if err := recover(); err != nil {
+			klog.Error(err)
+		}
+	}()
+
+	var persistentVolumeSource core_v1.PersistentVolumeSource
+	if strings.Contains(args.assetPath, ":") {
+		persistentVolumeSource = getNfsPersistentVolumeSource(args.assetPath)
+	} else {
+		persistentVolumeSource = getHostPathPersistentVolumeSource(args.assetPath)
+	}
 
 	return &core_v1.PersistentVolume{
 		TypeMeta: meta_v1.TypeMeta{
@@ -85,12 +137,7 @@ func getHostPathPV(args *pvArguments) *core_v1.PersistentVolume {
 				Namespace: args.pvc.Namespace,
 				UID:       args.pvc.UID,
 			},
-			PersistentVolumeSource: core_v1.PersistentVolumeSource{
-				HostPath: &core_v1.HostPathVolumeSource{
-					Type: hostPathType,
-					Path: args.assetPath,
-				},
-			},
+			PersistentVolumeSource: persistentVolumeSource,
 		},
 		Status: core_v1.PersistentVolumeStatus{},
 	}